remove LibChat
