--- vscode/webviews/LibChat.module.css
@@ -1,79 +1,0 @@
-.inner-container {
-    display: flex;
-    flex-direction: column;
-    height: 100%;
-}
-
-.transcript-container {
-    flex: 1;
-}
-
-.input-row {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    gap: 0.25rem;
-    padding: 0.5rem;
-}
-
-.text-area-container {
-    position: relative;
-    width: 100%;
-    display: flex;
-    align-items: flex-end;
-}
-
-.chat-input-container {
-    width: 100%;
-    height: 100%;
-    position: relative;
-}
-
-.chat-input {
-    width: 100%;
-    height: 100%;
-    resize: none;
-}
-
-.submit-button {
-    border: none;
-    cursor: pointer;
-
-    border-radius: 50%;
-    height: 2rem;
-    width: 2rem;
-
-    margin-left: 0.5rem;
-}
-
-.context-button {
-    color: var(--button-primary-background);
-    position: absolute;
-    right: 0.3rem;
-    bottom: 0.3rem;
-    fill: currentColor;
-    background: none;
-    border: none;
-    cursor: pointer;
-}
-
-.suggestions {
-    display: flex;
-    flex-wrap: wrap;
-    flex-direction: row;
-    gap: 0.1rem;
-    margin-bottom: 0.25rem;
-    /* This matches the button paddings, so the button text lines up with the input */
-    margin-left: -6px;
-}
-
-.abort-button-container {
-    display: flex;
-    /* Remove the padding of the form container from the width */
-    width: calc(100% - 2rem);
-    justify-content: center;
-    align-items: center;
-    position: absolute;
-    top: -2.5rem;
-    z-index: 100;
-}
