+++ vscode/webviews/chat/Transcript.tsx
@@ -4,15 +4,13 @@ import classNames from 'classnames'
 
 import type { ChatMessage, Guardrails, ModelProvider } from '@sourcegraph/cody-shared'
 
-import type {
-    ApiPostMessage,
-    ChatButtonProps,
-    ChatModelDropdownMenuProps,
-    CodeBlockActionsProps,
-    EditButtonProps,
-    FeedbackButtonsProps,
-    UserAccountInfo,
-} from '../LibChat'
+import type { UserAccountInfo } from '../Chat'
+import type { ChatButtonProps } from '../Chat'
+import type { EditButtonProps } from '../Chat'
+import type { FeedbackButtonsProps } from '../Chat'
+import type { ApiPostMessage } from '../Chat'
+import type { ChatModelDropdownMenuProps } from '../Components/ChatModelDropdownMenu'
+import type { CodeBlockActionsProps } from './CodeBlocks'
 
 import type { SymbolLinkProps } from './PreciseContext'
 import { TranscriptItem, type TranscriptItemClassNames } from './TranscriptItem'
